refactoring territorymap componnets. IslandInfoBox,LoadingIndicator,MapView,PlayerInfo
